# src/openapi/openapi.yml
openapi: 3.0.3
info:
  title: CB-SAU — Sistema de Autenticación de Usuarios
  version: 1.0.0
  description: >
    Microservicio SAU de CiberShield Colombia.
    Gestiona registro, autenticación JWT, perfil de empresa y activos digitales.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/auth/password-reset:
    post:
      tags: [Auth]
      summary: Solicitar restablecimiento de contraseña
      description: Envía un email con enlace de recuperación. Endpoint público.
      operationId: auth.passwordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Email de recuperación enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Si el correo existe, recibirá instrucciones de recuperación
        '400':
          description: Email inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Profile ────────────────────────────────────────────────────────────────
...